Move cli code to separate file and export compile function
